Refactor utils/auxiliars.py and utils/security.py into security/auth.py
