--- agent/core_agent/utils/security.py
@@ -1,66 +1,0 @@
-import logging
-
-import google.auth
-import google.oauth2.id_token
-from google.auth.transport.requests import Request
-from google.adk.agents.readonly_context import ReadonlyContext
-
-logger = logging.getLogger(__name__)
-
-
-def get_id_token(audience: str) -> str | None:
-    """
-    Generate a valid ID token to call a GCP service such as Cloud Run.
-
-    It first tries the metadata server (when running on GCP) and then falls back to
-    local ADC credentials for development.
-    """
-    request = Request()
-    try:
-        logger.debug(
-            "Retrieving ID token from metadata server for audience %s", audience
-        )
-        id_token = google.oauth2.id_token.fetch_id_token(request, audience)
-        logger.debug("ID token successfully retrieved from metadata server")
-        return id_token
-    except Exception as exc:
-        logging.debug("Metadata-server ID token retrieval failed: %s", exc)
-
-    try:
-        logging.debug("Retrieving ID token from local ADC credentials")
-        credentials, _ = google.auth.default()
-        credentials.refresh(request)
-        id_token = getattr(credentials, "id_token", None)
-        if id_token:
-            logging.debug("ID token retrieved from local ADC credentials")
-            return id_token
-        logging.warning("ADC credentials did not yield an ID token")
-    except Exception as exc:
-        logging.warning("Unable to obtain ID token from local ADC credentials: %s", exc)
-
-    return None
-
-
-def get_ge_oauth_token(readonly_context: ReadonlyContext, auth_id: str) -> str | None:
-    """
-    Inject the generated OAuth token that is generated by Gemini Enterprise when a Resource ID
-    is created.
-
-    Code adapted from: https://github.com/google/adk-docs/issues/1001#issuecomment-3894834825
-
-    Args:
-        readonly_context: The readonly context of the agent.
-        auth_id: The ID of the resource id in Gemini Enterprise.
-
-    Returns:
-        str | None: The OAuth token if found, None otherwise.
-    """
-    logger.info(f"Getting OAuth token for {auth_id = }")
-    oauth_token = readonly_context.state.get(auth_id)
-    if oauth_token:
-        logger.info("OAuth token found")
-    else:
-        logger.error("OAuth token not found")
-        logger.error(f"Available keys: {readonly_context.state.keys()}")
-
-    return oauth_token
